Modificações para entrega final lab03

+++ Aplicacao/Backend/src/main/java/com/puc/moedaestudantil/dto/StudentProfileResponse.java
@@ -2,10 +2,6 @@ package com.puc.moedaestudantil.dto;
 
 import io.micronaut.serde.annotation.Serdeable;
 
-/**
- * StudentProfileResponse - DTO para exibição dos dados do perfil do aluno
- * Utilizada no endpoint GET /api/students/{id}/profile
- */
 @Serdeable
 public record StudentProfileResponse(
     Long id,
@@ -17,6 +13,6 @@ public record StudentProfileResponse(
     String curso,
     Long instituicaoId,
     String instituicaoNome,
-    Long saldoMoedas
+    Integer saldoMoedas
 ) {
 }
